test(tidy): let clang-tidy analyse the tests, and clear what it finds
Invoke-ClangTidy.ps1 -IncludeTests could not complete. The tests leave fields
out of designated initializers, which is valid C++ that MSVC accepts; clang
warns about it, and /WX in the compile database made the warning an error.
It also passed the fuzz targets, which only the fuzz preset configures, so
clang-tidy guessed at their flags and failed to compile them. So the test
code had never been held to clang-tidy.

tests/.clang-tidy now turns that one warning off for test files through
ExtraArgs, and adds readability-function-cognitive-complexity to its list
of checks that only misfire here: Catch2's macros expand into branches. The
script analyses only files the compile database knows, naming any it skips.

The 50 findings that were left are fixed rather than silenced: const where it
can be, designated initializers, reserve before a loop, a named parameter,
cmp_equal and cmp_less_equal, any_of, CHECK_THROWS_AS for an empty catch,
unused using-declarations, and fixture methods that used nothing of the
fixture made free functions. Test constants whose construction allocates
(the emoji refs, the mirror map) are function-local statics, and one is a
string_view. Two are marked: command_info's implicit move, which can throw
because MSVC's std::map move allocates and only tests move one, and a toggle
test whose variable changes through a pointer to function the const check
cannot follow. -IncludeTests is clean.

+++ tests/unit/log_test.cpp
@@ -78,7 +78,7 @@ TEST_CASE("a message is never split between threads", "[log][threads]") {
     }
 
     const auto lines = captured.lines();
-    REQUIRE(lines.size() == thread_count * per_thread);
+    REQUIRE(lines.size() == static_cast<std::size_t>(thread_count) * per_thread);
     // Every line has to be one complete message: interleaving would leave
     // fragments that do not match the format.
     for (const auto& [level, message] : lines) {
